Unify watcher management
